# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Label Dvisha e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      responses:
        "200":
          description: All categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    patch:
      operationId: updateCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── PRODUCTS ──────────────────────────────────────────────────────────────
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "201":
          description: Added to cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      parameters:
        - name: approved
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Gallery images
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryListResponse"
    post:
      operationId: uploadGallery
      tags: [gallery]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryInput"
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
  /gallery/{id}/approve:
    patch:
      operationId: approveGallery
      tags: [gallery, admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
  /gallery/{id}:
    delete:
      operationId: deleteGallery
      tags: [gallery, admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── UPLOAD (Cloudinary signature) ─────────────────────────────────────────
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, slug, sortOrder, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        cloudinaryPublicId:
          type: ["string", "null"]
        sortOrder:
          type: integer
        isActive:
          type: boolean
    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        cloudinaryPublicId:
          type: string
        sortOrder:
          type: integer
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        cloudinaryPublicId:
          type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean

    # Products
...
    Product:
      type: object
      required: [id, name, slug, sku, price, isActive, isFeatured, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        shortDescription:
          type: ["string", "null"]
        sku:
          type: string
        price:
          type: number
        compareAtPrice:
          type: ["number", "null"]
        categoryId:
          type: ["integer", "null"]
        material:
          type: ["string", "null"]
        styleTag:
          type: ["string", "null"]
        deliveryDays:
          type: ["string", "null"]
        allowCustomSize:
          type: boolean
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        lowStockThreshold:
          type: integer
        createdAt:
          type: string
        primaryImage:
          type: ["string", "null"]
        category:
          $ref: "#/components/schemas/Category"
...
    CustomMeasurements:
      type: object
      properties:
        bust:
          type: string
        waist:
          type: string
        hip:
          type: string
        height:
          type: string
    CartItem:
      type: object
      required: [id, cartId, productId, quantity]
      properties:
        id:
          type: integer
        cartId:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        colorId:
          type: ["integer", "null"]
        colorName:
          type: ["string", "null"]
        productSizeId:
          type: ["integer", "null"]
        sizeLabel:
          type: ["string", "null"]
        customMeasurements:
          $ref: "#/components/schemas/CustomMeasurements"
        quantity:
          type: integer
    Cart:
      type: object
      required: [id, items, itemCount, subtotal]
      properties:
        id:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        itemCount:
          type: integer
        subtotal:
          type: number
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        colorId:
          type: integer
        productSizeId:
          type: integer
        customMeasurements:
          $ref: "#/components/schemas/CustomMeasurements"
        quantity:
          type: integer
    CartItemUpdate:
      type: object
      properties:
        quantity:
          type: integer
        colorId:
          type: integer
        productSizeId:
          type: integer
        customMeasurements:
          $ref: "#/components/schemas/CustomMeasurements"

    # Orders
...
    GalleryItem:
      type: object
      required: [id, userId, imageUrl, isApproved, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        imageUrl:
          type: string
        cloudinaryPublicId:
          type: ["string", "null"]
        caption:
          type: ["string", "null"]
        isApproved:
          type: boolean
        createdAt:
          type: string
    GalleryInput:
      type: object
      required: [imageUrl]
      properties:
        imageUrl:
          type: string
        cloudinaryPublicId:
          type: string
        caption:
          type: string
    GalleryListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/GalleryItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Upload
...