# lib/api-zod/src/generated/types/galleryInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Label Dvisha e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryInput {
  imageUrl: string;
  cloudinaryPublicId?: string;
  caption?: string;
}